Changed how pulling works, and error checking
Added error checking to the opening of the file. I then changed how the grabbing for the files worked, allowing for the multiple files with the same class to be called and the contents to be taken out. Also fixed the formatting for the file so that it can be read easier by the injector program